Core/Maps: move GridMap out of Map.h/cpp
