+++ src/common/collision/Maps/MapDefines.cpp
@@ -1,0 +1,24 @@
+/*
+* This file is part of the TrinityCore Project. See AUTHORS file for Copyright information
+*
+* This program is free software; you can redistribute it and/or modify it
+* under the terms of the GNU General Public License as published by the
+* Free Software Foundation; either version 2 of the License, or (at your
+* option) any later version.
+*
+* This program is distributed in the hope that it will be useful, but WITHOUT
+* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+* FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+* more details.
+*
+* You should have received a copy of the GNU General Public License along
+* with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include "MapDefines.h"
+
+u_map_magic const MapMagic        = { {'M','A','P','S'} };
+u_map_magic const MapVersionMagic = { {'v','1','.','9'} };
+u_map_magic const MapAreaMagic    = { {'A','R','E','A'} };
+u_map_magic const MapHeightMagic  = { {'M','H','G','T'} };
+u_map_magic const MapLiquidMagic  = { {'M','L','I','Q'} };
